+++ tests/Js2IL.Test262.Tests/language/statements/do-while/ExecutionTests.cs
@@ -41,4 +41,17 @@ public class ExecutionTests : FileSystemExecutionTestsBase
     public Task S12_6_1_A4_T4()
         => ExecutionTest("S12.6.1_A4_T4");
 
+
+
+    [Fact(DisplayName = "S12.6.1_A2", Skip = "ReferenceError propagation from do-while bodies is not yet aligned with this upstream case.")]
+    public Task S12_6_1_A2()
+        => ExecutionTest("S12.6.1_A2");
+
+    [Fact(DisplayName = "S12.6.1_A4_T5")]
+    public Task S12_6_1_A4_T5()
+        => ExecutionTest("S12.6.1_A4_T5");
+
+    [Fact(DisplayName = "S12.6.1_A9", Skip = "ReferenceError propagation from do-while bodies is not yet aligned with this upstream case.")]
+    public Task S12_6_1_A9()
+        => ExecutionTest("S12.6.1_A9");
 }